cleanup

+++ ypom/tweetnacl.c
@@ -2,6 +2,8 @@
 #define FOR(i,n) for (i = 0;i < n;++i)
 #define sv static void
 
+
+
 typedef unsigned char u8;
 typedef unsigned long u32;
 typedef unsigned long long u64;
@@ -282,7 +284,7 @@ sv car25519(gf o)
     }
 }
 
-sv sel25519(gf p,gf q,int b)
+sv sel25519(gf p,gf q,i64 b)
 {
     i64 t,i,c=~(b-1);
     FOR(i,16) {
@@ -516,7 +518,7 @@ static const u64 K[80] =
     0x4cc5d4becb3e42b6ULL, 0x597f299cfc657e2aULL, 0x5fcb6fab3ad6faecULL, 0x6c44198c4a475817ULL
 };
 
-int crypto_hashblocks(u8 *x,const u8 *m,u64 n)
+u64 crypto_hashblocks(u8 *x,const u8 *m,u64 n)
 {
     u64 z[8],b[8],a[8],w[16],t;
     int i,j;
